requirements.txt
# Core Data Science & ML
pandas==2.0.0
numpy==1.23.5
scikit-learn==1.3.0
scipy==1.11.0

# API & Data Fetching
requests==2.31.0
python-dotenv==1.0.0

# Web Framework
flask==2.3.0
flask-cors==4.0.0

# Serialization & Storage
joblib==1.3.0
pickle-mixin==1.0

# Monitoring & Logging
sentry-sdk==1.32.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.10.0
flake8==6.1.0

# Notebooks & Visualization
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0

# Deployment (optional)
gunicorn==21.2.0
python-decouple==3.8

# Documentation (optional)
sphinx==7.2.0
sphinx-rtd-theme==1.3.0
